Restructing of JS/Web (split); updated docs, see especially stack.rst
