+++ exa/_config.py
@@ -62,14 +62,14 @@ _conf['logfile_max_count'] = 5
 
 # Internal package paths
 pkg = os.path.dirname(__file__)
-_conf['static_templates'] = mkp(pkg, 'templates')
-_conf['static_html'] = mkp(pkg, 'static', 'html')
-_conf['static_css'] = mkp(pkg, 'static', 'css')
-_conf['static_js'] = mkp(pkg, 'static', 'js')
+_conf['web_templates'] = mkp(pkg, 'web', 'templates')
+_conf['web_html'] = mkp(pkg, 'web', 'html')
+_conf['web_css'] = mkp(pkg, 'web', 'css')
+_conf['web_js'] = mkp(pkg, 'web', 'js')
 _conf['static_constants.json'] = mkp(pkg, 'static', 'constants.json')
 _conf['static_isotopes.json'] = mkp(pkg, 'static', 'isotopes.json')
 _conf['static_units.json'] = mkp(pkg, 'static', 'units.json')
-_conf['nbext_localdir'] = mkp(pkg, 'static', 'nbextensions')
+_conf['nbext_localdir'] = mkp(pkg, 'nbextensions')
 _conf['nbext_sysdir'] = mkp(jupyter_data_dir(), 'nbextensions', 'exa')
