+++ exa/nbextensions/apps/gui.js
@@ -9,7 +9,7 @@ Basic gui for container views.
 
 require.config({
     shim: {
-        'nbextensions/exa/lib/dat.gui.min': {
+        'nbextensions/exa/libs/dat.gui.min': {
             exports: 'dat'
         },
     },
@@ -17,7 +17,7 @@ require.config({
 
 
 define([
-    'nbextensions/exa/lib/dat.gui.min',
+    'nbextensions/exa/libs/dat.gui.min',
 ], function(dat) {
     class ContainerGUI extends dat.GUI {
         /*"""
