--- exa/static/nbextensions/__paths__
@@ -1,3 +1,0 @@
-'js/lib/three.min.js'
-'js/lib/TrackballControls.js'
-'js/tester.js'
