--- exa/static/nbextensions/tester.js
@@ -1,1 +1,0 @@
-../js/tester.js
