+++ exa/web/__init__.py
@@ -1,0 +1,1 @@
+from exa.web.main import serve
